# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QnA API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDtoPaged"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDtoPaged"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDtoPaged"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/mark-read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/users/{id}/stats": {
      "get": {
        "tags": [
          "Stats"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/overview": {
      "get": {
        "tags": [
          "Stats"
        ],
        "parameters": [
          {
            "name": "top",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewStatsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewStatsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewStatsDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "MarkReadRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "isRead": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "actorId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "actorName": {
            "type": "string",
            "nullable": true
          },
          "postId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "postTitle": {
            "type": "string",
            "nullable": true
          },
          "commentId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "excerpt": {
            "type": "string",
            "nullable": true
          },
          "actionUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NotificationDtoPaged": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OverviewStatsDto": {
        "type": "object",
        "properties": {
          "totalQuestions": {
            "type": "integer",
            "format": "int64"
          },
          "totalComments": {
            "type": "integer",
            "format": "int64"
          },
          "totalVotes": {
            "type": "integer",
            "format": "int64"
          },
          "totalUsers": {
            "type": "integer",
            "format": "int64"
          },
          "topTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopTagDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TopTagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UserStatsDto": {
        "type": "object",
        "properties": {
          "questions": {
            "type": "integer",
            "format": "int32"
          },
          "comments": {
            "type": "integer",
            "format": "int32"
          },
          "votes": {
            "type": "integer",
            "format": "int32"
          },
          "reputation": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Nhập token theo dạng: Bearer {token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}